refactor: splice test code, reuse add, remove className code
