+++ packages/toast-ui.grid/src/helper/common.ts
@@ -147,7 +147,9 @@ export function assign(targetObj: Obj, obj: Obj) {
 
 export function removeArrayItem<T>(targetItem: T, arr: T[]) {
   const targetIdx = findIndex(item => item === targetItem, arr);
-  arr.splice(targetIdx, 1);
+  if (targetIdx !== -1) {
+    arr.splice(targetIdx, 1);
+  }
 
   return arr;
 }
