+++ packages/toast-ui.grid/src/query/column.ts
@@ -1,5 +1,5 @@
 import { ComplexColumnInfo, Column } from '../store/types';
-import { some, uniq } from '../helper/common';
+import { some } from '../helper/common';
 
 export function isParentColumnHeader(complexColumnHeaders: ComplexColumnInfo[], name: string) {
   return !!complexColumnHeaders.length && some(item => item.name === name, complexColumnHeaders);
@@ -7,25 +7,4 @@ export function isParentColumnHeader(complexColumnHeaders: ComplexColumnInfo[], 
 
 export function isHiddenColumn(column: Column, columnName: string) {
   return column.allColumnMap[columnName].hidden;
-}
-
-export function addClassName(className: string, prevClassNames?: string[]) {
-  const classNames = className.split(' ');
-  const columnClassNames = prevClassNames ? prevClassNames : [];
-
-  return uniq([...classNames, ...columnClassNames]);
-}
-
-export function removeClassName(className: string, prevClassNames: string[]) {
-  const classNames = className.split(' ');
-  const removedClassNames = prevClassNames;
-
-  classNames.forEach(clsName => {
-    const index = removedClassNames.indexOf(clsName);
-    if (index !== -1) {
-      removedClassNames.splice(index, 1);
-    }
-  });
-
-  return removedClassNames;
 }
